requirements.txt
ackermann_msgs==2.0.2
ament_index_python==1.4.0
controller==2023.1.3
cv_bridge==3.2.1
fastseg==0.1.2
geometry_msgs==4.2.4
launch_ros==0.19.8
message_filters==4.3.5
nav2_common==1.1.18
nav2_msgs==1.1.17
nav_msgs==4.2.4
numpy==2.3.1
opencv_python==4.10.0.84
Pillow==11.2.1
PyYAML==5.4.1
PyYAML==6.0.2
rclpy==3.3.15
rosgraph_msgs==1.2.1
sensor_msgs==4.2.4
setuptools==75.6.0
setuptools==59.6.0
std_msgs==4.2.4
tf2_ros_py==0.25.9
torch==1.13.1
ultralytics==8.3.38
webots_ros2_driver==2023.1.3
